refactor: reservation domain 패키지 내부로 이동
